o god

+++ code/game/objects/effects/effect_system/effects_smoke.dm
@@ -50,7 +50,7 @@
 	if(lifetime < 1)
 		kill_smoke()
 		return FALSE
-	for(var/mob/living/L in range(0,src))
+	for(var/mob/living/L in get_turf(src))
 		smoke_mob(L)
 	return TRUE
 
@@ -165,32 +165,32 @@
 	var/weldvents = TRUE
 	var/distcheck = TRUE
 
-/datum/effect_system/smoke_spread/freezing/proc/Chilled(atom/A)
-	if(isopenturf(A))
-		var/turf/open/T = A
-		if(T.air)
-			var/datum/gas_mixture/G = T.air
-			if(!distcheck || get_dist(T, location) < blast) // Otherwise we'll get silliness like people using Nanofrost to kill people through walls with cold air
-				G.temperature = temperature
+/datum/effect_system/smoke_spread/freezing/proc/Chilled(turf/open/T)
+	if(!istype(T))
+		return
+	if(T.air)
+		var/datum/gas_mixture/G = T.air
+		if(!distcheck || get_dist(T, location) < blast) // Otherwise we'll get silliness like people using Nanofrost to kill people through walls with cold air
+			G.temperature = temperature
 			T.air_update_turf(FALSE, FALSE)
-			for(var/obj/effect/hotspot/H in T)
-				qdel(H)
-			var/list/G_gases = G.gases
-			if(G_gases[/datum/gas/plasma])
-				G.assert_gas(/datum/gas/nitrogen)
-				G_gases[/datum/gas/nitrogen][MOLES] += (G_gases[/datum/gas/plasma][MOLES])
-				G_gases[/datum/gas/plasma][MOLES] = 0
-				G.garbage_collect()
-		if (weldvents)
-			for(var/obj/machinery/atmospherics/components/unary/U in T)
-				if(!isnull(U.welded) && !U.welded) //must be an unwelded vent pump or vent scrubber.
-					U.welded = TRUE
-					U.update_appearance()
-					U.visible_message(span_danger("[U] is frozen shut!"))
-		for(var/mob/living/L in T)
-			L.extinguish_mob()
-		for(var/obj/item/Item in T)
-			Item.extinguish()
+		for(var/obj/effect/hotspot/H in T)
+			qdel(H)
+		var/list/G_gases = G.gases
+		if(G_gases[/datum/gas/plasma])
+			G.assert_gas(/datum/gas/nitrogen)
+			G_gases[/datum/gas/nitrogen][MOLES] += (G_gases[/datum/gas/plasma][MOLES])
+			G_gases[/datum/gas/plasma][MOLES] = 0
+			G.garbage_collect()
+	if (weldvents)
+		for(var/obj/machinery/atmospherics/components/unary/U in T)
+			if(!isnull(U.welded) && !U.welded) //must be an unwelded vent pump or vent scrubber.
+				U.welded = TRUE
+				U.update_appearance()
+				U.visible_message(span_danger("[U] is frozen shut!"))
+	for(var/mob/living/L in T)
+		L.extinguish_mob()
+	for(var/obj/item/Item in T)
+		Item.extinguish()
 
 /datum/effect_system/smoke_spread/freezing/set_up(radius = 5, loca, blast_radius = 0)
 	..()
@@ -198,7 +198,7 @@
 
 /datum/effect_system/smoke_spread/freezing/start()
 	if(blast)
-		for(var/turf/T in RANGE_TURFS(blast, location))
+		for(var/turf/open/T in RANGE_TURFS(blast, location))
 			Chilled(T)
 	..()
